@@ -5,25 +5,29 @@ declare(strict_types=1);
 namespace Synolia\SyliusSchedulerCommandPlugin\Checker;
 
 use Cron\CronExpression;
+use Symfony\Component\DependencyInjection\Attribute\AutoconfigureTag;
 use Synolia\SyliusSchedulerCommandPlugin\Components\Exceptions\Checker\IsNotDueException;
 use Synolia\SyliusSchedulerCommandPlugin\Entity\CommandInterface;
 
 /**
  * This checker only works if current date/time is checked every minutes
  */
+#[AutoconfigureTag(IsDueCheckerInterface::class)]
 class EveryMinuteIsDueChecker implements IsDueCheckerInterface
 {
+    private const PRIORITY = 0;
+
     public static function getDefaultPriority(): int
     {
-        return 0;
+        return self::PRIORITY;
     }
 
     /**
-     * @throws \Synolia\SyliusSchedulerCommandPlugin\Components\Exceptions\Checker\IsNotDueException
+     * @throws IsNotDueException
      */
     public function isDue(CommandInterface $command, ?\DateTimeInterface $dateTime = null): bool
     {
-        if (null === $dateTime) {
+        if (!$dateTime instanceof \DateTimeInterface) {
             $dateTime = new \DateTime();
         }
 
