@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace Synolia\SyliusSchedulerCommandPlugin\Controller;
 
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\DependencyInjection\Attribute\Autowire;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Contracts\Translation\TranslatorInterface;
 use Synolia\SyliusSchedulerCommandPlugin\DataRetriever\LogDataRetriever;
 use Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryInterface;
@@ -15,26 +17,29 @@ use Synolia\SyliusSchedulerCommandPlugin\Repository\ScheduledCommandRepositoryIn
 final class LogViewerController extends AbstractController
 {
     public function __construct(
-        private ScheduledCommandRepositoryInterface $scheduledCommandRepository,
-        private LogDataRetriever $logDataRetriever,
-        private TranslatorInterface $translator,
-        private string $logsDir,
+        private readonly ScheduledCommandRepositoryInterface $scheduledCommandRepository,
+        private readonly LogDataRetriever $logDataRetriever,
+        private readonly TranslatorInterface $translator,
+        #[Autowire(param: 'kernel.logs_dir')]
+        private readonly string $logsDir,
         /** @var int The time in milliseconds between two AJAX requests to the server. */
-        private int $updateTime = 2000,
+        private readonly int $updateTime = 2000,
     ) {
     }
 
+    #[Route('/scheduled-commands/{command}/get-log.json', name: 'sylius_admin_scheduler_get_log_file', defaults: ['_sylius' => ['permission' => true]], methods: ['GET'])]
     public function getLogs(Request $request, string $command): JsonResponse
     {
         $scheduleCommand = $this->scheduledCommandRepository->find($command);
 
-        if (null === $scheduleCommand ||
+        if (
+            null === $scheduleCommand ||
             null === $scheduleCommand->getLogFile()
         ) {
             return new JsonResponse('', Response::HTTP_NO_CONTENT);
         }
 
-        if (true === (bool) $request->get('refresh')) {
+        if ((bool) $request->get('refresh')) {
             $result = $this->logDataRetriever->getLog(
                 $this->logsDir . \DIRECTORY_SEPARATOR . $scheduleCommand->getLogFile(),
                 (int) $request->get('lastsize'),
@@ -56,11 +61,13 @@ final class LogViewerController extends AbstractController
         ]);
     }
 
+    #[Route('/scheduled-commands/{command}/view-log', name: 'sylius_admin_scheduler_view_log_file', defaults: ['_sylius' => ['permission' => true]], methods: ['GET'])]
     public function show(string $command): Response
     {
         $scheduledCommand = $this->scheduledCommandRepository->find($command);
 
-        if (null === $scheduledCommand ||
+        if (
+            null === $scheduledCommand ||
             null === $scheduledCommand->getLogFile()
         ) {
             $this->addFlash('error', $this->translator->trans('sylius.ui.does_not_exists_or_missing_log_file'));
