@@ -16,11 +16,11 @@ final class SynoliaSyliusSchedulerCommandExtension extends Extension implements 
     use PrependDoctrineMigrationsTrait;
 
     /**
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
+     * @inheritDoc
      */
-    public function load(array $config, ContainerBuilder $container): void
+    public function load(array $configs, ContainerBuilder $container): void
     {
-        $loader = new YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
+        $loader = new YamlFileLoader($container, new FileLocator(\dirname(__DIR__, 2) . '/config'));
         $loader->load('services.yaml');
     }
 
@@ -39,7 +39,7 @@ final class SynoliaSyliusSchedulerCommandExtension extends Extension implements 
 
     protected function getMigrationsDirectory(): string
     {
-        return '@SynoliaSyliusSchedulerCommandPlugin/Migrations';
+        return '@SynoliaSyliusSchedulerCommandPlugin/migrations';
     }
 
     protected function getNamespacesOfMigrationsExecutedBefore(): array
