@@ -4,22 +4,23 @@ declare(strict_types=1);
 
 namespace Synolia\SyliusSchedulerCommandPlugin\Voter;
 
+use Symfony\Component\DependencyInjection\Attribute\TaggedIterator;
+use Synolia\SyliusSchedulerCommandPlugin\Checker\IsDueCheckerInterface;
 use Synolia\SyliusSchedulerCommandPlugin\Components\Exceptions\Checker\IsNotDueException;
 use Synolia\SyliusSchedulerCommandPlugin\Entity\CommandInterface;
 
 class IsDueVoter implements IsDueVoterInterface
 {
-    /** @var array<\Synolia\SyliusSchedulerCommandPlugin\Checker\IsDueCheckerInterface> */
-    private array $isDueCheckers;
-
-    public function __construct(\Traversable $checkers)
-    {
-        $this->isDueCheckers = iterator_to_array($checkers);
+    public function __construct(
+        #[TaggedIterator(IsDueCheckerInterface::class)]
+        private readonly iterable $checkers,
+    ) {
     }
 
     public function isDue(CommandInterface $command, ?\DateTimeInterface $dateTime = null): bool
     {
-        foreach ($this->isDueCheckers as $checker) {
+        /** @var IsDueCheckerInterface $checker */
+        foreach ($this->checkers as $checker) {
             try {
                 return $checker->isDue($command, $dateTime);
             } catch (IsNotDueException) {
